+++ packages/builder/builder-rspack-provider/src/core/createCompiler.ts
@@ -1,13 +1,11 @@
 import {
   debug,
   logger,
+  prettyTime,
   formatStats,
   TARGET_ID_MAP,
-  getProgressColor,
 } from '@modern-js/builder-shared';
 import type { Context, RspackConfig } from '../types';
-import chalk from '@modern-js/utils/chalk';
-import prettyTime from '@modern-js/builder-shared/pretty-time';
 
 export async function createCompiler({
   context,
@@ -37,13 +35,12 @@ export async function createCompiler({
     if (!stats.hasErrors()) {
       obj.children?.forEach((c, index) => {
         if (c.time) {
-          const color = chalk[getProgressColor(index)];
-          const time = prettyTime([0, c.time * 10 ** 6], 1);
+          const time = prettyTime([0, c.time * 10 ** 6]);
           const target = Array.isArray(context.target)
             ? context.target[index]
             : context.target;
           const name = TARGET_ID_MAP[target || 'web'];
-          logger.log(color(`✔ ${name}  compiled in`, time));
+          logger.ready(`${name} compiled in ${time}`);
         }
       });
     }