+++ packages/builder/builder-webpack-provider/src/webpackPlugins/ProgressPlugin/ProgressPlugin.ts
@@ -1,7 +1,6 @@
 import webpack from 'webpack';
 import { logger } from '@modern-js/utils/logger';
-import prettyTime from '@modern-js/builder-shared/pretty-time';
-import ForkTsCheckerWebpackPlugin from '@modern-js/builder-shared/fork-ts-checker-webpack-plugin';
+import { prettyTime } from '@modern-js/builder-shared';
 import { bus, createFriendlyPercentage } from './helpers';
 import { createNonTTYLogger } from './helpers/nonTty';
 import type { Props } from './helpers/type';
@@ -10,30 +9,22 @@ export interface ProgressOptions
   extends Omit<Partial<Props>, 'message' | 'total' | 'current' | 'done'> {
   id?: string;
   clearOnDone?: boolean;
-  showRecompileLog?: boolean;
   disableTsChecker?: boolean;
 }
 
 export class ProgressPlugin extends webpack.ProgressPlugin {
   readonly name: string = 'ProgressPlugin';
 
-  hasTypeErrors: boolean = false;
+  id: string;
 
   hasCompileErrors: boolean = false;
 
   compileTime: string | null = null;
 
-  showRecompileLog: boolean;
-
   disableTsChecker?: boolean;
 
   constructor(options: ProgressOptions) {
-    const {
-      id = 'Modern',
-      clearOnDone = false,
-      showRecompileLog = false,
-      disableTsChecker,
-    } = options;
+    const { id = 'Modern', clearOnDone = false, disableTsChecker } = options;
 
     const nonTTYLogger = createNonTTYLogger();
     const friendlyPercentage = createFriendlyPercentage();
@@ -59,7 +50,6 @@ export class ProgressPlugin extends webpack.ProgressPlugin {
             message,
             done,
             hasErrors: this.hasCompileErrors,
-            compileTime: this.compileTime,
           });
           bus.render();
 
@@ -78,7 +68,7 @@ export class ProgressPlugin extends webpack.ProgressPlugin {
       },
     });
 
-    this.showRecompileLog = showRecompileLog;
+    this.id = id;
     this.disableTsChecker = disableTsChecker;
   }
 
@@ -86,45 +76,22 @@ export class ProgressPlugin extends webpack.ProgressPlugin {
     super.apply(compiler);
 
     let startTime: [number, number] | null = null;
-    let isReCompile = false;
 
     compiler.hooks.compile.tap(this.name, () => {
-      // If it is a recompile and there are compilation errors,
-      // print a recompile log so that users can know that the recompile
-      // is triggered and the above error log is outdated.
-      if (
-        isReCompile &&
-        this.showRecompileLog &&
-        (this.hasCompileErrors || this.hasTypeErrors)
-      ) {
-        logger.info(`Start recompile...\n`);
-      }
-
       this.compileTime = null;
       startTime = process.hrtime();
-      isReCompile = true;
     });
 
     compiler.hooks.done.tap(this.name, stat => {
       if (startTime) {
         this.hasCompileErrors = stat.hasErrors();
-        this.compileTime = prettyTime(process.hrtime(startTime), 1);
+        this.compileTime = prettyTime(process.hrtime(startTime));
         startTime = null;
+
+        if (!this.hasCompileErrors) {
+          logger.ready(`${this.id} compiled in ${this.compileTime}`);
+        }
       }
     });
-
-    if (!this.disableTsChecker) {
-      const hooks = ForkTsCheckerWebpackPlugin.getCompilerHooks(compiler);
-
-      hooks.start.tap(this.name, change => {
-        this.hasTypeErrors = false;
-        return change;
-      });
-
-      hooks.issues.tap(this.name, issues => {
-        this.hasTypeErrors = Boolean(issues.length);
-        return issues;
-      });
-    }
   }
 }