+++ packages/devtools/plugin/src/modern-app-env.d.ts
@@ -1,1 +1,3 @@
 /// <reference types='@modern-js/types/hoist-non-react-statics' />
+
+declare let __resourceQuery: string;