+++ packages/server/plugin-express/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @modern-js/plugin-express
 
+## 2.35.1
+
+### Patch Changes
+
+- Updated dependencies [ea3fe18]
+- Updated dependencies [9dd3151]
+- Updated dependencies [4980480]
+- Updated dependencies [6a1d46e]
+  - @modern-js/utils@2.35.1
+  - @modern-js/bff-core@2.35.1
+  - @modern-js/bff-runtime@2.35.1
+  - @modern-js/types@2.35.1
+
 ## 2.35.0
 
 ### Patch Changes