+++ packages/solutions/module-tools/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @modern-js/module-tools
 
+## 2.35.1
+
+### Patch Changes
+
+- ce24aad: feat: bump postcss-nesting v12 to support latest CSS nesting syntax
+
+  feat: 升级 postcss-nesting v12，支持最新的 CSS nesting 语法
+
+- Updated dependencies [ea3fe18]
+- Updated dependencies [a45cc5b]
+- Updated dependencies [14f95cf]
+- Updated dependencies [4b8cdd3]
+- Updated dependencies [9dd3151]
+- Updated dependencies [4980480]
+- Updated dependencies [14b0906]
+- Updated dependencies [843cb2c]
+- Updated dependencies [6a1d46e]
+  - @modern-js/utils@2.35.1
+  - @modern-js/libuild-plugin-swc@2.35.1
+  - @modern-js/plugin@2.35.1
+  - @modern-js/core@2.35.1
+  - @modern-js/plugin-changeset@2.35.1
+  - @modern-js/plugin-i18n@2.35.1
+  - @modern-js/plugin-lint@2.35.1
+  - @modern-js/new-action@2.35.1
+  - @modern-js/upgrade@2.35.1
+  - @modern-js/libuild-plugin-svgr@2.35.1
+  - @modern-js/libuild@2.35.1
+  - @modern-js/types@2.35.1
+
 ## 2.35.0
 
 ### Patch Changes